requirements.txt
pyserial
docopt
types-docopt
pytest
